added logic to comment on pull requests
