+++ src/listeners/pullRequestListener.ts
@@ -1,12 +1,10 @@
 import { Probot } from "probot";
-import configs from "../configs/github.webhook.event.configs.json";
 
-export function listeningForPullRequestEvents(app: Probot) {
+export function listeningForPullRequestEvents(app: Probot, events: any[]) {
   return new Promise((resolve, reject) => {
     try {
       app.log.info("Listening for pull request events");
 
-      const events: any[] = [configs.pull_request.opened];
       app.on(events, async (context) => {
         app.log.info("Received an pull request event");
