+++ src/main.ts
@@ -2,12 +2,14 @@ import { Probot } from "probot";
 import { connectDb } from "./db/dbConnection";
 import { listeningForAppInstallationEvents } from "./listeners/appInstallationListener";
 import { fetchDetailsWithInstallationId } from "./fetch/fetchBase";
-import { processRepositories } from "./processing/processRepositories";
+import { processRepositories } from "./services/repositoryService";
 import { listeningForPullRequestEvents } from "./listeners/pullRequestListener";
-import { getAllFilesFromPullRequest } from "./fetch/getAllFiles";
 import configs from "./configs/fetch.configs.json";
-import { GithubResponseFile } from "./types/GithubResponseFile";
-import { File } from "./db/models/File";
+import eventConfigs from "./configs/github.webhook.event.configs.json";
+import { createCommentOnGithub } from "./services/commentService";
+import { processPullRequests } from "./services/pullRequestService";
+import { FileScoreMap } from "./types/FileScoreMap";
+import { markdownTable } from "markdown-table";
 
 export async function main(app: Probot) {
   connectDb(app).catch((error: any) => {
@@ -42,7 +44,13 @@ export async function main(app: Probot) {
         status: file.status,
       }));
 
-      // app.log.info(files);
+      // const comment = files.map((file: ResponseFile) => {
+      //   return markdownTable([
+      //     ["File Name", "Status"],
+      //     [`${file.filePath}`, `${file.status}`],
+      //   ]);
+      // });
+      // app.log.info(comment);
     })
     .catch((error: any) => {
       app.log.error(error);
@@ -50,7 +58,9 @@ export async function main(app: Probot) {
 }
 
 function processAppInstallationEvents(app: Probot) {
-  listeningForAppInstallationEvents(app)
+  const events: any[] = [eventConfigs.app_installation.created];
+
+  listeningForAppInstallationEvents(app, events)
     .then((response: any) => processRepositories(app, response))
     .catch((error: any) => {
       app.log.error("Error while processing app installation event");
@@ -59,66 +69,19 @@ function processAppInstallationEvents(app: Probot) {
 }
 
 function processPullRequestEvents(app: Probot) {
-  listeningForPullRequestEvents(app)
-    .then((context: any) => processPullRequests(app, context.payload))
+  const events: any[] = [eventConfigs.pull_request.opened];
+
+  listeningForPullRequestEvents(app, events)
+    .then(async (context: any) => {
+      const files: FileScoreMap[] = await processPullRequests(
+        app,
+        context.payload
+      );
+      const comment = `File name ${files[0].fileName} with score: ${files[0].score}`;
+      createCommentOnGithub(comment, context);
+    })
     .catch((error: any) => {
       app.log.error("Error while processing pull request events");
       app.log.error(error);
     });
-}
-
-async function processPullRequests(app: Probot, response: any) {
-  const { pull_request, installation } = response;
-  const pullNumber: number = pull_request.number;
-
-  // todo: thinking how to use them
-  // const mainBranchRef: string = pull_request.base.ref;
-  // const mainBranchSha: string = pull_request.base.sha;
-  const repoFullName: string = pull_request.base.repo.full_name;
-  const installationId: number = installation.id;
-
-  const fullName = repoFullName.split("/");
-  const owner = fullName[0];
-  const repoName = fullName[1];
-
-  const responseFiles: GithubResponseFile[] = await getAllFilesFromPullRequest(
-    app,
-    owner,
-    repoName,
-    installationId,
-    pullNumber
-  );
-
-  type FileScoreMap = {
-    fileName: string;
-    score: number;
-  };
-
-  const fileScoreMap: FileScoreMap[] = await Promise.all(
-    responseFiles.map(async (file: GithubResponseFile, index) => {
-      const scores: number[] | null = await File.findOne(
-        {
-          installationId: installationId,
-          owner: owner,
-          repoName: repoName,
-          filePath: file.filePath,
-        },
-        "riskScore"
-      );
-
-      if (scores == null) {
-        return {
-          fileName: file.filePath,
-          score: 0,
-        };
-      }
-
-      return {
-        fileName: file.filePath,
-        score: scores[index],
-      };
-    })
-  );
-
-  return fileScoreMap;
 }
