+++ src/types/FileScoreMap.ts
@@ -1,0 +1,4 @@
+export type FileScoreMap = {
+  fileName: string;
+  score: number;
+};
